✨ refactor imports to use useAuthValue hook from hooks directory
