+++ src/context/AuthContext.jsx
@@ -2,7 +2,7 @@
 
 import { createContext } from "react"
 
-const AuthContext = createContext()
+export const AuthContext = createContext()
 
 export function AuthProvider(props){
     const { children, value } = props
